support links as well as real files as filenames
My music collection is in a git-annex so my collection is
basically a big link farm playd ignored as it checks for the
existence of a real file. The test is factored out and internally
split into an existence check and check for real or link now.
This also means we have to check for directories first before
checking for file existence and the find command to list files in
a given directory must be told to follow links