avoid playback of the first file before stepping to last position
There is a delay between loading of playlist and skipping to the
last position in the playlist (it was even enforced with a sleep?)
so we put mplayer into 'pause' before loading playlist and skipping
and end the pause after it.

Also implements prefix-support (for e.g. 'pausing' …) in playd_put